% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{finisterra_server_connect}
\alias{finisterra_server_connect}
\title{Connect to FINISTERRA Database}
\usage{
finisterra_server_connect(username = NULL)
}
\arguments{
\item{username}{Your database username}
}
\value{
A connection object
}
\description{
Connect to FINISTERRA Database
}
