% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{get_site_data}
\alias{get_site_data}
\title{Get Site Data}
\usage{
get_site_data(con, db)
}
\arguments{
\item{con}{Connection object from finisterra_server_connect()}

\item{db}{The database name ('esc', 'cari', 'gdc')}
}
\value{
Dataframe with joined Context and XYZ tables from database
}
\description{
Get Site Data
}
