% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{get_site_datums}
\alias{get_site_datums}
\title{Get Site Datums}
\usage{
get_site_datums(con, db)
}
\arguments{
\item{con}{Connection object from finisterra_server_connect()}

\item{db}{The database name ('esc', 'cari', 'gdc')}
}
\value{
Dataframe with Datums table from the database
}
\description{
Get Site Datums
}
