% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{get_site_xyz}
\alias{get_site_xyz}
\title{Get Site XYZ}
\usage{
get_site_xyz(con, site)
}
\arguments{
\item{con}{Connection object from finisterra_server_connect()}

\item{site}{The abbreviated site name ('esc', 'cari', 'gdc')}
}
\value{
Dataframe with Context table from database
}
\description{
Get Site XYZ
}
