% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{search_by_square}
\alias{search_by_square}
\title{Search by square
Search for all XYZ records from a specific excavation unit}
\usage{
search_by_square(con, site, unit_id)
}
\arguments{
\item{con}{Connection object from finisterra_server_connect()}

\item{site}{The abbreviated site name ('esc', 'cari', 'gdc')}

\item{unit_id}{The name of the unit of interest}
}
\value{
Dataframe with joined Context and XYZ tables from specific unit
}
\description{
Search by square
Search for all XYZ records from a specific excavation unit
}
