% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{get_site_datums}
\alias{get_site_datums}
\title{Get Site Datums}
\usage{
get_site_datums(con, site)

get_site_datums(con, site)
}
\arguments{
\item{con}{Connection object from finisterra_server_connect()}

\item{site}{The abbreviated site name ('esc', 'cari', 'gdc')}
}
\value{
Dataframe with Datums table from the database

Dataframe with Units table from the database
}
\description{
Get Site Datums

Get Site Units
}
